requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
kafka-python==2.3.0
cassandra-driver==3.29.3
redis==5.0.1
asyncpg==0.30.0
PyYAML==6.0.1
pydantic==2.12.5
langchain-google-genai==3.0.1
langchain-core==1.0.3